remove old unused PlayersTab, rename new one "PlayersTab"
